ai-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pillow==12.1.1
numpy==1.26.4; python_version < "3.13"
numpy==2.2.4; python_version >= "3.13"
scikit-learn==1.5.1; python_version < "3.13"
scikit-learn==1.6.1; python_version >= "3.13"
joblib==1.4.2
